remove parseLayer2MetadataForTransferV1
