+++ packages/shared/lib/core/layer-2/utils/index.ts
@@ -7,8 +7,6 @@ export * from './getLayer2AssetAllowance'
 export * from './getLayer2MetadataForTransfer'
 export * from './getLayer2NetworkFromAddress'
 export * from './parseLayer2Metadata'
-export * from './parseLayer2MetadataForTransferV1'
-export * from './parseLayer2MetadataForTransferV2'
 export * from './parseLayer2MetadataForTransfer'
 export * from './specialNativeTokenAmountEncoding'
 export * from './outputHexBytes'
